+++ common/src/main/java/dev/terminalmc/clientsort/client/gui/widget/TransferButton.java
@@ -74,7 +74,7 @@ public class TransferButton extends TriggerButton {
                 isPlayerInv,
                 isPlayerInv ? SPRITES_UP : SPRITES_DOWN,
                 name,
-                policy == null ? null : policy.className(),
+                policy == null ? null : policy.getKey(),
                 lowestPolicyKey,
                 offset,
                 policy == null ? true : policy.offsetFromSlot(),
@@ -107,7 +107,7 @@ public class TransferButton extends TriggerButton {
         if (policy == null) {
             options().classPolicies.put(
                     key,
-                    new ClassPolicy(
+                    ClassPolicy.create(
                             key,
                             offset,
                             offsetFromSlot,
@@ -125,7 +125,7 @@ public class TransferButton extends TriggerButton {
         } else {
             options().classPolicies.put(
                     key,
-                    new ClassPolicy(
+                    ClassPolicy.create(
                             key,
                             offset,
                             offsetFromSlot,